opengl getters

+++ include/libmeme/Common.hpp
@@ -92,7 +92,7 @@
 // misc
 #define _ML						::ml::
 #define ML_alias				using // global typedef
-#define ML_addressof(ptr)		((void *)(intmax_t)(ptr))
+#define ML_addressof(ptr)		((void *)(intptr_t)(ptr))
 #define ML_arraysize(arr)		(sizeof(arr) / sizeof(*arr))
 #define ML_compare(lhs, rhs)	(((lhs) != (rhs)) ? (((lhs) < (rhs)) ? -1 : 1) : 0)
 #define ML_forward(x)			std::forward<decltype(x)>(x)
